azure_functions/requirements.txt
azure-functions-core-tools
azure-functions
